+++ src/components/Checkbox.tsx
@@ -78,10 +78,13 @@ export const Checkbox = <Inputs extends Record<string, unknown>>({
       {!!tooltipContent && (
         <div className="flex items-center p-3">
           <Popover className="relative">
-            <PopoverButton className="flex items-center justify-center w-6 h-6 text-slate-400 hover:text-slate-600 transition-colors duration-200">
+            <PopoverButton
+              aria-label="Show token details"
+              className="flex items-center justify-center w-6 h-6 text-slate-400 hover:text-slate-600 transition-colors duration-200"
+            >
               <InformationCircleIcon className="w-6 h-6" />
             </PopoverButton>
-            <PopoverPanel className="absolute z-10 w-100 px-4 mt-3 transform -translate-x-1/2 left-1/2 sm:px-0">
+            <PopoverPanel className="absolute z-10 px-4 mt-3 transform -translate-x-1/2 left-1/2 sm:px-0">
               <div className="overflow-hidden rounded-lg shadow-lg ring-1 ring-black ring-opacity-5">
                 <div className="relative bg-white p-4 text-sm text-slate-700">
                   {tooltipContent}
